# floramatch-terraform/functions/garden-service/garden.routes.js
const { plantsCollection } = require("./firestore");
const { getUserId } = require("./auth");

function send(res, code, body) {
  res.status(code).json(body);
}

exports.handleGardenRoutes = async (req, res) => {
  try {

    console.log("AUTH HEADER:", req.headers.authorization);

    res.set('Access-Control-Allow-Origin', '*');
    res.set('Access-Control-Allow-Methods', 'GET, POST, PUT, DELETE, OPTIONS');
    res.set('Access-Control-Allow-Headers', 'Content-Type, Authorization');

    if (req.method === 'OPTIONS') {
      return res.status(204).send('');
    }

    console.log("HEADERS:", JSON.stringify(req.headers, null, 2));


    const userId = await getUserId(req);
    const col = plantsCollection(userId);
    const path = req.path;
    const method = req.method;

    if (method === "GET" && path === "/garden/plants") {
      const snapshot = await col.get();
      const plants = snapshot.docs.map(d => ({
        plantInstanceId: d.id,
        ...d.data()
      }));
      return send(res, 200, plants);
    }

    if (method === "GET" && path.startsWith("/garden/plants/")) {
      const id = path.split("/").pop();
      const doc = await col.doc(id).get();

      if (!doc.exists) {
        return send(res, 404, { error: "Plant not found" });
      }

      return send(res, 200, {
        plantInstanceId: doc.id,
        ...doc.data()
      });
    }

    // POPRAWIONA WERSJA (Zapisuje wszystkie parametry roślinki)
    if (method === "POST" && path === "/garden/plants") {
      const body = req.body;

      if (!body.plantId) {
        return send(res, 400, { error: "plantId required" });
      }

      const doc = await col.add({
        plantId: body.plantId,
        nickname: body.nickname || "",
        location: body.location || "",
        purchaseDate: body.purchaseDate || null,
        notes: body.notes || "",
        locationType: body.locationType || "",
        sunlight: body.sunlight || "",
        watering: body.watering || "",
        plantSize: body.plantSize || "",
        flowering: body.flowering || "",
        ...body, // Bezpiecznie dołącza wszystkie pozostałe przesłane właściwości
        createdAt: new Date().toISOString()
      });

      return send(res, 201, {
        status: "success",
        message: "Plant added to garden",
        plantInstanceId: doc.id
      });
    }

    if (method === "PUT" && path.startsWith("/garden/plants/")) {
      const id = path.split("/").pop();
      const docRef = col.doc(id);
      const doc = await docRef.get();

      if (!doc.exists) {
        return send(res, 404, { error: "Plant not found" });
      }

      await docRef.update({
        ...req.body,
        updatedAt: new Date().toISOString()
      });

      return send(res, 200, { 
        status: "success", 
        message: "Plant updated successfully" 
      });
    }

    if (method === "DELETE" && path.startsWith("/garden/plants/")) {
      const id = path.split("/").pop();
      const docRef = col.doc(id);
      const doc = await docRef.get();

      if (!doc.exists) {
        return send(res, 404, { error: "Plant not found" });
      }

      await docRef.delete();
      return send(res, 200, { 
        status: "success", 
        message: "Plant removed from garden" 
      });
    }

    return send(res, 404, { error: "Not Found" });

  } catch (e) {
    return send(res, 401, { error: e.message });
  }
};