# floramatch-terraform/openapi.yaml
swagger: "2.0"

info:
  title: FloraMatch API
  version: 1.0.0

schemes:
  - https

produces:
  - application/json

x-google-endpoints:
  - name: floramatch-gateway-4rnl9enj.ew.gateway.dev
    allowCors: true

securityDefinitions:
  firebase:
    type: oauth2
    flow: implicit
    authorizationUrl: ""

    x-google-issuer: https://securetoken.google.com/floramatch-497314

    x-google-jwks_uri: https://www.googleapis.com/service_accounts/v1/jwk/securetoken@system.gserviceaccount.com

    x-google-audiences: floramatch-497314

#security:
#  - firebase: []

paths:

  /recommendations:

    options:
      security: []
      operationId: recommendationsOptions

      x-google-backend:
        address: https://plant-recommendation-service-o4dtgtgq7a-lm.a.run.app
        jwt_audience: https://plant-recommendation-service-o4dtgtgq7a-lm.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS

      responses:
        "204":
          description: CORS preflight

    post:
      operationId: getRecommendations
      security:
        - firebase: [ ]
      x-google-backend:
        address: https://plant-recommendation-service-o4dtgtgq7a-lm.a.run.app

        jwt_audience: https://plant-recommendation-service-o4dtgtgq7a-lm.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "200":
          description: OK

  /garden/plants:

    options:
      security: []
      operationId: gardenPlantsOptions

      x-google-backend:
        address: https://garden-service-o4dtgtgq7a-lm.a.run.app
        jwt_audience: https://garden-service-o4dtgtgq7a-lm.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS

      responses:
        "204":
          description: CORS preflight

    get:
      operationId: getPlants
      security:
        - firebase: [ ]
      x-google-backend:
        address: https://garden-service-o4dtgtgq7a-lm.a.run.app

        jwt_audience: https://garden-service-o4dtgtgq7a-lm.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "200":
          description: OK

    post:
      operationId: addPlant
      security:
        - firebase: [ ]
      x-google-backend:
        address: https://garden-service-o4dtgtgq7a-lm.a.run.app

        jwt_audience: https://garden-service-o4dtgtgq7a-lm.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "200":
          description: OK

  /garden/plants/{id}:

    options:
      security: []
      operationId: gardenPlantByIdOptions

      parameters:
        - name: id
          in: path
          required: true
          type: string

      x-google-backend:
        address: https://garden-service-o4dtgtgq7a-lm.a.run.app
        jwt_audience: https://garden-service-o4dtgtgq7a-lm.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS

      responses:
        "204":
          description: CORS preflight

    get:
      security:
        - firebase: [ ]
      operationId: getPlantById
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://garden-service-o4dtgtgq7a-lm.a.run.app

        jwt_audience: https://garden-service-o4dtgtgq7a-lm.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "200":
          description: OK

    put:
      security:
        - firebase: [ ]
      operationId: updatePlant
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://garden-service-o4dtgtgq7a-lm.a.run.app

        jwt_audience: https://garden-service-o4dtgtgq7a-lm.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "200":
          description: OK

    delete:
      security:
        - firebase: [ ]
      operationId: deletePlant
      parameters:
        - name: id
          in: path
          required: true
          type: string
      x-google-backend:
        address: https://garden-service-o4dtgtgq7a-lm.a.run.app

        jwt_audience: https://garden-service-o4dtgtgq7a-lm.a.run.app
        path_translation: APPEND_PATH_TO_ADDRESS
      responses:
        "200":
          description: OK